manual cli argument parsing; remove golf
I have intended to manually parse command line arguments for a while,
and decided to give it a go. While more verbose, definitely has more
flexibility, and can serve as an example for other projects I'm
working on.

Also, maybe the key is to take the lessons learned here, and throw
them back into golf, so it can be more flexible.